feat:splitting on modules
