+++ index.html
@@ -234,6 +234,6 @@
       </section>
     </template>
 <script src="D:\Keks-repo\2607777-kekstagram-2\js\functions.js"></script>
-<script src="D:\Keks-repo\2607777-kekstagram-2\js\main.js"></script>
+<script src="D:\Keks-repo\2607777-kekstagram-2\js\main.js" type="module"></script>
 </body>
 </html>
